Moved test-ISA types (Inst, Block, Func) into tests and implemented Function trait. Regallocs still go through direct interface though (TODO).
